Crate swf_headers

A library for reading data from the header of a .swf file

Structs

DecodedSwf
Handles decompressing swf innards and reading the results.

SwfHeaders
A struct containing all the parsed headers of a swf file.

Enums

Error
The error type used by swf-headers.

Signature
An enum representing all the valid signatures of a swf file

Traits

BitRange
Helper trait for getting a range of bits from a container of bytes as a u32.