Enum swf_headers::Signature

pub enum Signature {
    Uncompressed,
    ZlibCompressed,
    LzmaCompressed,
}

An enum representing all the valid signatures of a SWF file.

As of the latest SWF specification, there are 3 valid signatures a SWF file can have. The first three bytes of a SWF file act as the magic numbers, FWS (SWF backwards) was defined with the original specification, and designates an uncompressed SWF file. CWS was introduced with SWF 6, and specifies that all bytes beyond the first 8 are compressed using zlib. ZWS was added with SWF 13, and displays the same concept, but with LZMA instead of zlib.

Variants

Uncompressed

A signature of FWS, meaning an uncompressed SWF file.

ZlibCompressed

A signature of CWS, meaning a zlib-compressed SWF file.

LzmaCompressed

A signature of ZWS, meaning an LZMA-compressed SWF file.