Trait swf_headers::BitRange

pub trait BitRange {
    fn get_bit_range(&self, range: Range<u32>) -> u32;
}

Helper trait for getting a range of bits from a container of bytes as a u32.

Required Methods

fn get_bit_range(&self, range: Range<u32>) -> u32

Takes a range and converts the bits in that range into a u32.

Example: rust let vec = vec![0b0010_1100, 0b0111_0010]; assert!(vec.get_bit_range(2..12) == 0b1011000111);