Crate swf_headers

A library for reading data from the header of a .swf file that can also be used as the ground-works for a larger library that works on parsing the rest.

Github repo

Crates.io

Parsing SWF files can be tricky, which is why I made this crate to help you out! Not only does it provide a convenient way to read through the headers of any SWF file, but it also gives you a readable stream of everything just after the header, with all compression issues sorted out for you! How useful.

Structs

DecodedSwf
Handles decompressing swf innards and reading the results.

SwfHeaders
The primary struct, managing all the parsing and storage of SWF header information.

Enums

Error
The error type used by swf-headers.

Signature
An enum representing all the valid signatures of a SWF file.